UnoCore: add EGL 1.0 wrappers
Android uses EGL, and it can often be useful to have access to the
various EGL objects. So let's expose EGL more or less as a whole.

The only function missing is eglGetProcAddress, which kind of
fundamentally doesn't work, due to Unos inability to grok function
pointers. Something like Marshal.GetDelegateForFunctionPointer would
be needed to support this.